+++ src/Lynx/Search/MoveOrdering.cs
@@ -173,7 +173,7 @@ public sealed partial class Engine
             int rawHistoryMalus = HistoryMalus[depth];
 
             ref var quietHistoryEntry = ref QuietHistoryEntry(position, move, ref evaluationContext);
-            quietHistoryEntry = ScoreHistoryMove(quietHistoryEntry, rawHistoryBonus);
+            quietHistoryEntry = (short)ScoreHistoryMove(quietHistoryEntry, rawHistoryBonus);
 
             if (!isRoot)
             {
@@ -197,7 +197,7 @@ public sealed partial class Engine
                     // 🔍 Quiet history penalty / malus
                     // When a quiet move fails high, penalize previous visited quiet moves
                     quietHistoryEntry = ref QuietHistoryEntry(position, visitedMove, ref evaluationContext);
-                    quietHistoryEntry = ScoreHistoryMove(quietHistoryEntry, -rawHistoryMalus);
+                    quietHistoryEntry = (short)ScoreHistoryMove(quietHistoryEntry, -rawHistoryMalus);
 
                     if (!isRoot)
                     {
@@ -242,7 +242,7 @@ public sealed partial class Engine
         var rawHistoryMalus = HistoryMalus[depth];
 
         ref var captureHistoryEntry = ref CaptureHistoryEntry(move);
-        captureHistoryEntry = ScoreHistoryMove(captureHistoryEntry, rawHistoryBonus);
+        captureHistoryEntry = (short)ScoreHistoryMove(captureHistoryEntry, rawHistoryBonus);
 
         // 🔍 Capture history penalty/malus
         // When a capture fails high, penalize previous visited captures
@@ -255,7 +255,7 @@ public sealed partial class Engine
             if (capturedPiece != (int)Piece.None)
             {
                 ref var captureHistoryVisitedMove = ref CaptureHistoryEntry(visitedMove);
-                captureHistoryVisitedMove = ScoreHistoryMove(captureHistoryVisitedMove, -rawHistoryMalus);
+                captureHistoryVisitedMove = (short)ScoreHistoryMove(captureHistoryVisitedMove, -rawHistoryMalus);
             }
         }
     }
